% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trim_times_do.R
\name{assign_trim_end_do}
\alias{assign_trim_end_do}
\title{Create variable for DO test end time in UTC}
\usage{
assign_trim_end_do(Log)
}
\arguments{
\item{Log}{Metadata log with deployment_datetime column}
}
\value{
Returns DOendtime_utc
}
\description{
Create variable for DO test end time in UTC
}
\author{
Nicole Torrie
}
