% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trim_times_hdo.R
\name{assign_trim_end_hdo}
\alias{assign_trim_end_hdo}
\title{Create variable for HOBO DO test end time in UTC}
\usage{
assign_trim_end_hdo(log)
}
\arguments{
\item{log}{Metadata log with deployment_datetime column}
}
\value{
Returns hdo_endtime_utc
}
\description{
Create variable for HOBO DO test end time in UTC
}
\author{
Nicole Torrie
}
