% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trim_times_sal.R
\name{assign_trim_end_sal}
\alias{assign_trim_end_sal}
\title{Create variable for SAL test end time in UTC}
\usage{
assign_trim_end_sal(Log)
}
\arguments{
\item{Log}{Metadata log with deployment_datetime column}
}
\value{
Returns SALendtime_utc
}
\description{
Create variable for SAL test end time in UTC
}
\author{
Nicole Torrie
}
