% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trim_times_temp.R
\name{assign_trim_end_temp}
\alias{assign_trim_end_temp}
\title{Create variable for TEMP test end time in UTC}
\usage{
assign_trim_end_temp(log)
}
\arguments{
\item{log}{Metadata log with deployment_datetime column}
}
\value{
Returns temp_endtime_utc
}
\description{
Create variable for TEMP test end time in UTC
}
\author{
Nicole Torrie
}
