% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trim_times_hdo.R
\name{assign_trim_start_hdo}
\alias{assign_trim_start_hdo}
\title{Create variable for Hobo DO test start time in UTC}
\usage{
assign_trim_start_hdo(log)
}
\arguments{
\item{log}{Metadata log with deployment_datetime column}
}
\value{
Returns hdo_starttime_utc
}
\description{
Create variable for Hobo DO test start time in UTC
}
\author{
Nicole Torrie
}
