% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trim_times_all.R
\name{assign_trim_times_all}
\alias{assign_trim_times_all}
\title{Create table of test start and end times for each variable}
\usage{
assign_trim_times_all(var_list, log)
}
\arguments{
\item{var_list}{list of variables present in the dataset}

\item{log}{Metadata log with deployment_datetime and retrieval_datetime columns}
}
\value{
a data frame of all test start and end times in utc
}
\description{
Create table of test start and end times for each variable
}
\author{
Nicole Torrie
}
