% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_log_all.R
\name{clean_log_all}
\alias{clean_log_all}
\title{Clean log for export}
\usage{
clean_log_all(Log = Log)
}
\arguments{
\item{Log}{Metadata log}
}
\value{
Returns cleaned log which sensorstrings can import
}
\description{
Clean log for export
}
\author{
Nicole Torrie
}
