% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_log_duplicates.R
\name{clean_log_duplicates}
\alias{clean_log_duplicates}
\title{Clean duplicate rows out of metadata log}
\usage{
clean_log_duplicates(Log)
}
\arguments{
\item{Log}{Metadata log}
}
\value{
Returns cleaned log with no duplicate rows from multi-variable sensors
}
\description{
Clean duplicate rows out of metadata log
}
\author{
Nicole Torrie
}
