% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_val_log.R
\name{create_val_log}
\alias{create_val_log}
\title{Create metadata log from Tracking Google sheet}
\usage{
create_val_log(Tracking)
}
\arguments{
\item{Tracking}{Dataframe created from downloaded CalVal Tracking Google sheet}
}
\value{
Returns metadata log with updated column names and new columns with deployment and retrieval datetime in Atlantic timezone
}
\description{
Create metadata log from Tracking Google sheet
}
\author{
Nicole Torrie
}
