% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_assign_tolerance_flag.R
\name{cv_assign_tolerance_flag}
\alias{cv_assign_tolerance_flag}
\title{Assign a flag to each observation}
\usage{
cv_assign_tolerance_flag(dat, tolerance = NULL, vr2ar_tol = 0.5)
}
\arguments{
\item{dat}{Data frame of compiled validation data with column
\code{round_timestamp}, as returned from \code{cv_round_timestamps()}.}

\item{tolerance}{A data frame with the specified accuracy for each variable.
Must have two columns: \code{variable}, which has an entry for each
variable in \code{dat}, and \code{tol}, with the accuracy. Note this
assumes only one type of sensor for each variable, or that the sensors have
the same tolerance. Exception: the tolerance for the VR2AR temperature
sensors may be set in \code{vr2ar_tol} argument.}

\item{vr2ar_tol}{The tolerance for the VR2AR temperature sensors. Set
\code{vr2ar_tol = NULL} to use the same value for all temperature sensors.}
}
\value{
Returns dat with several additional columns: tolerance, med,
  tol_lower, tol_upper, and qc_flag.
}
\description{
The tolerance flag is based on the median value of the time interval
plus/minus the specified accuracy for each variable.
}
