% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_create_val_log.R
\name{cv_create_val_log}
\alias{cv_create_val_log}
\title{Create metadata log from calval tracking sheet}
\usage{
cv_create_val_log(path, tracking)
}
\arguments{
\item{path}{File path to the Log folder.}

\item{tracking}{calval tracking sheet, as downloaded from
\code{cv_read_calval_tracking()}.}
}
\value{
Exports csv file to /log folder in the log format required for
  compiling data with \code{sensorstrings}.
}
\description{
Create metadata log from calval tracking sheet
}
