% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cv_import_path}
\alias{cv_import_path}
\title{Generate file path to import raw calval data}
\usage{
cv_import_path(val_id, path = NULL, check_path = TRUE)
}
\arguments{
\item{val_id}{Character string of the validation id. Will be converted to
lower case, and all spaces will be replaced with an underscore.}

\item{path}{File path to the validation_data folder on the CMAR R drive.}

\item{check_path}{Logical parameter indicating whether to check if the path
exists. Default is /code{TRUE}.}
}
\value{
The file path for importing raw calval data.
}
\description{
Raw data must be saved in a folder path/validation_data/valXXX or
path/validation_data/postXXX.
}
