% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_round_timestamps.R
\name{cv_round_timestamps}
\alias{cv_round_timestamps}
\title{Round timestamps to given intervals}
\usage{
cv_round_timestamps(dat, round_ints = NULL)
}
\arguments{
\item{dat}{Data frame of compiled validation data in a long format.}

\item{round_ints}{Data frame indicating what intervals to use when rounding
the timestamps for each variable. Must have two columns: \code{variable},
which has an entry for each variable in \code{dat}, and \code{round_ints},
a character string of the corresponding interval, e.g., "10 minutes". The
default interval is 10 minutes for dissolved oxygen and salinity, and 15
minutes for temperature.}
}
\value{
Returns dat with an additional column \code{round_timestamp}.
}
\description{
Round timestamps to given intervals
}
