% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_trim_dat.R
\name{cv_trim_dat}
\alias{cv_trim_dat}
\title{Trim compiled data to the start and end of each test}
\usage{
cv_trim_dat(dat, trimtimes)
}
\arguments{
\item{dat}{Data frame of compiled validation data in wide format.}

\item{trimtimes}{Data frame with the start and end timestamps for each
variable. Must have three columns: variable, deployment_utc, retrieval_utc.}
}
\value{
Returns dat in a long format, trimmed to the start and end of the
  test for each variable.
}
\description{
Trim compiled data to the start and end of each test
}
