% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_val_trimtimes.R
\name{cv_val_trimtimes}
\alias{cv_val_trimtimes}
\title{Create table of validation test start and end times for each variable}
\usage{
cv_val_trimtimes(tracking)
}
\arguments{
\item{tracking}{calval tracking sheet, as downloaded from
\code{cv_read_calval_tracking()}.}
}
\value{
Returns a data frame of the start and end times for each variable, in
  UTC.
}
\description{
Create table of validation test start and end times for each variable
}
