% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_do_flag.R
\name{ggplot_do_flag}
\alias{ggplot_do_flag}
\title{Plot DO validation data, colorize points by DO flag}
\usage{
ggplot_do_flag(DO, point_size = 0.75)
}
\arguments{
\item{DO}{Dataframe filtered for dissolved oxygen percent saturation test data}

\item{point_size}{to set size of ggplot points}
}
\value{
Returns a ggplot object of Dissolved Oxygen validation data colorized based on DO flag
}
\description{
Plot DO validation data, colorize points by DO flag
}
\author{
Nicole Torrie
}
