% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_do_val.R
\name{ggplot_do_val}
\alias{ggplot_do_val}
\title{Plot DO validation data, colorize points by sensor serial #}
\usage{
ggplot_do_val(DO, point_size = 0.75)
}
\arguments{
\item{DO}{Dataframe filtered for dissolved oxygen percent saturation test data}

\item{point_size}{to set size of ggplot points}
}
\value{
Returns a ggplot object of Dissolved Oxygen validation data colorized based on sensor serial #
}
\description{
Plot DO validation data, colorize points by sensor serial #
}
\author{
Nicole Torrie
}
