% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_hdo_flag.R
\name{ggplot_hdo_flag}
\alias{ggplot_hdo_flag}
\title{Plot HOBO DO validation data, colorize points by HOBO DO flag}
\usage{
ggplot_hdo_flag(final_HDO)
}
\arguments{
\item{final_HDO}{Dataframe filtered for HDO test data}
}
\value{
Returns a ggplot object of HOBO DO validation data colorized based on HOBO DO flag
}
\description{
Plot HOBO DO validation data, colorize points by HOBO DO flag
}
\author{
Nicole Torrie
}
