% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sal_flag.R
\name{ggplot_sal_flag}
\alias{ggplot_sal_flag}
\title{Plot SAL validation data, colorize points by SAL flag}
\usage{
ggplot_sal_flag(final_SAL, point_size = 0.75)
}
\arguments{
\item{final_SAL}{Dataframe filtered for salinity test data}

\item{point_size}{to set size of ggplot points}
}
\value{
Returns a ggplot object of Salinity validation data colorized based on SAL flag
}
\description{
Plot SAL validation data, colorize points by SAL flag
}
\author{
Nicole Torrie
}
