% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sal_val.R
\name{ggplot_sal_val}
\alias{ggplot_sal_val}
\title{Plot SAL validation data, colorize points by sensor serial #}
\usage{
ggplot_sal_val(final_SAL, point_size = 0.75)
}
\arguments{
\item{final_SAL}{Dataframe filtered for salinity test data}

\item{point_size}{to set size of ggplot points}
}
\value{
Returns a ggplot object of Salinity validation data colorized based on sensor serial #
}
\description{
Plot SAL validation data, colorize points by sensor serial #
}
\author{
Nicole Torrie
}
