% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_temp_flag.R
\name{ggplot_temp_flag}
\alias{ggplot_temp_flag}
\title{Plot TEMP validation data, colorize points by Temp flag}
\usage{
ggplot_temp_flag(final_temp)
}
\arguments{
\item{final_temp}{Dataframe filtered for temperature test data}
}
\value{
Returns a ggplot object of Temperature validation data colorized based on temp flag
}
\description{
Plot TEMP validation data, colorize points by Temp flag
}
\author{
Nicole Torrie
}
