% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_log_dates.R
\name{unify_log_dates}
\alias{unify_log_dates}
\title{Unify dates in metadata log}
\usage{
unify_log_dates(Log = Log)
}
\arguments{
\item{Log}{Metadata log}
}
\value{
Returns cleaned log with unified deployment and retrieval dates for each sensor
}
\description{
Unify dates in metadata log
}
\author{
Nicole Torrie
}
