% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trim_times_hdo.R
\name{assign_trim_end_HDO}
\alias{assign_trim_end_HDO}
\title{Create variable for HOBO DO test end time in UTC}
\usage{
assign_trim_end_HDO(Log)
}
\arguments{
\item{Log}{Metadata log with deployment_datetime column}
}
\value{
Returns HDOendtime_utc
}
\description{
Create variable for HOBO DO test end time in UTC
}
\author{
Nicole Torrie
}
