% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_create_log.R
\name{cv_create_log}
\alias{cv_create_log}
\title{Create metadata log from calval tracking sheet}
\usage{
cv_create_log(event_id, path = NULL)
}
\arguments{
\item{event_id}{The event_id for which to create the log. Must match an entry
in the event_id column of the calval tracking sheet.}

\item{path}{File path to the Log folder. Default is the log folder in the
event_id folder on the CMAR R drive.}
}
\value{
Exports csv file to /log folder in the log format required for
  compiling data with \code{sensorstrings}.
}
\description{
Create metadata log from calval tracking sheet
}
