% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_flags.R
\name{cv_ggplot_flags}
\alias{cv_ggplot_flags}
\title{Create ggplot for one variable}
\usage{
cv_ggplot_flags(
  dat,
  var,
  colour_col = "qc_flag",
  pal = NULL,
  plotly_friendly = FALSE
)
}
\arguments{
\item{dat}{Data frame of flagged validation data in long format. Must include
a column named with the string "qc_flag".}

\item{var}{Caharacter string of the variable to plot.}

\item{colour_col}{Character string indicating the column to use to colour the
observations.}

\item{pal}{Colour palette assigned to the observations.}

\item{plotly_friendly}{Logical argument. If \code{TRUE}, the legend will be
plotted when \code{plotly::ggplotly} is called on \code{p}. Default is
\code{FALSE}, which makes the legend look better in a static figure.}
}
\value{
Returns a ggplot object; a figure of validation results coloured by
  flag for \code{var}.
}
\description{
Create ggplot for one variable
}
