% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot_flags.R
\name{cv_plot_flags}
\alias{cv_plot_flags}
\title{Plot sensor data coloured by flag value}
\usage{
cv_plot_flags(
  dat,
  vars = "all",
  colour_col = "qc_flag",
  pal = NULL,
  plotly_friendly = FALSE
)
}
\arguments{
\item{dat}{Data frame of flagged validation data in long or wide format.}

\item{vars}{Character vector of variables to plot. Default is \code{vars =
"all"}, which will make a plot for each recognized variable in \code{dat}.}

\item{colour_col}{Character string indicating the column to use to colour the
observations.}

\item{pal}{Vector of colours onto which \code{colour_col} will be mapped.
Default is c("chartreuse4", "#DB4325") when \code{colour_col = "qc_flag"},
and Dark2 from RColourBrewer otherwise.}

\item{plotly_friendly}{Logical argument. If \code{TRUE}, the legend will be
plotted when \code{plotly::ggplotly} is called on \code{p}. Default is
\code{FALSE}, which makes the legend look better in a static figure.}
}
\value{
Returns a list of ggplot objects; one figure for each variable in
  \code{vars}.
}
\description{
Plot sensor data coloured by flag value
}
