% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_read_calval_tracking.R
\name{cv_read_calval_tracking}
\alias{cv_read_calval_tracking}
\title{Read in CMAR calval tracking sheet}
\usage{
cv_read_calval_tracking(link = NULL, sheet = NULL)
}
\arguments{
\item{link}{Link to the calval tracking sheet on Google Drive. Default is the
CMAR tracking sheet.}

\item{sheet}{Character string with the name of the sheet to read in.}
}
\value{
Returns a data frame of the calval tracking sheet.
}
\description{
Read in CMAR calval tracking sheet
}
