% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_round_timestamps.R
\name{cv_round_timestamps}
\alias{cv_round_timestamps}
\title{Round timestamps to given intervals}
\usage{
cv_round_timestamps(
  dat,
  do_percent_sat_int = "10 minutes",
  do_mg_l_int = "15 minutes",
  ph_int = "10 minutes",
  sal_int = "10 minutes",
  temp_int = "15 minutes"
)
}
\arguments{
\item{dat}{Data frame of compiled validation data in a long format.}

\item{do_percent_sat_int}{Character string indicating the time interval for
binning the dissolved oxygen (percent saturation) timestamps. Default is
\code{"10 minutes"}.}

\item{do_mg_l_int}{Character string indicating the time interval for
binning the dissolved oxygen (mg/L) timestamps. Default is \code{"10
minutes"}.}

\item{ph_int}{Character string indicating the time interval for binning the
pH timestamps. Default is \code{"15 minutes"}.}

\item{sal_int}{Character string indicating the time interval for binning the
salinity timestamps. Default is \code{"10 mintues"}.}

\item{temp_int}{Character string indicating the time interval for binning the
temperature timestamps. Default is \code{"15 minutes"}.}
}
\value{
Returns dat with an additional column \code{round_timestamp}.
}
\description{
Round timestamps to given intervals
}
