% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_summarise_flags.R
\name{cv_summarise_flags}
\alias{cv_summarise_flags}
\title{Title}
\usage{
cv_summarise_flags(dat, wide = TRUE)
}
\arguments{
\item{dat}{Data frame of compiled and flagged validation data, as returned
from \code{cv_assign_tolerance_flag()}.}

\item{wide}{Logical argument indicating whether to return a wide table of
percentage values, or a long table including counts and percents.}
}
\value{
Returns a summary of observations that passed and failed for each
  variable and sensor.
}
\description{
Title
}
