% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_temp_val.R
\name{ggplot_temp_val}
\alias{ggplot_temp_val}
\title{Plot TEMP validation data, colorize points by sensor serial #}
\usage{
ggplot_temp_val(final_temp, size = 0.75)
}
\arguments{
\item{final_temp}{Dataframe filtered for temperature test data}

\item{size}{Argument to set size of ggplot points}
}
\value{
Returns a ggplot object of temperature validation data colorized based on sensor serial #
}
\description{
Plot TEMP validation data, colorize points by sensor serial #
}
\author{
Nicole Torrie
}
